[backend/frontend] hardcoded template logic in back
