--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/README.md
@@ -1,1 +1,0 @@
-TEMPORARY. LATER, WIDGETS WILL BE ON BACKEND SIDE.
