--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerCreationDate.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerCreationDate: TemplateWidget = {
-  name: 'Container Creation Date',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerCreationDate',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerCreationDate;
