--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerDescription.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerDescription: TemplateWidget = {
-  name: 'Container Description',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerDescription',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerDescription;
