--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerLabels.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerLabels: TemplateWidget = {
-  name: 'Container Labels',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerLabels',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerLabels;
