--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerMarkings.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerMarkings: TemplateWidget = {
-  name: 'Container Markings',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerMarkings',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerMarkings;
