--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerModificationDate.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerModificationDate: TemplateWidget = {
-  name: 'Container Modification Date',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerModificationDate',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerModificationDate;
