--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerName.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerName: TemplateWidget = {
-  name: 'Container Name',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerName',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerName;
