--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerObservables.widget.ts
@@ -1,25 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerObservables: TemplateWidget = {
-  name: 'Container Observables',
-  widget: {
-    type: 'list',
-    id: 'widgetContainerObservables',
-    perspective: 'entities',
-    dataSelection: [
-      {
-        perspective: 'entities',
-        filters: {
-          mode: 'and',
-          filters: [
-            { key: 'entity_type', values: ['Indicator'] },
-            { key: 'objects', values: ['CONTAINER_ID'] },
-          ],
-          filterGroups: [],
-        },
-      },
-    ],
-  },
-};
-
-export default widgetContainerObservables;
