--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/containerReferences.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetContainerReferences: TemplateWidget = {
-  name: 'Container References',
-  widget: {
-    type: 'attribute',
-    id: 'widgetContainerReferences',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetContainerReferences;
