--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/incidentPriority.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetIncidentPriority: TemplateWidget = {
-  name: 'Incident Priority',
-  widget: {
-    type: 'attribute',
-    id: 'widgetIncidentPriority',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetIncidentPriority;
