--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/incidentSeverity.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetIncidentSeverity: TemplateWidget = {
-  name: 'Incident Severity',
-  widget: {
-    type: 'attribute',
-    id: 'widgetIncidentSeverity',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetIncidentSeverity;
