--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/incidentTTP.widget.ts
@@ -1,25 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetIncidentTTP: TemplateWidget = {
-  name: 'Incident TTPs',
-  widget: {
-    type: 'list',
-    id: 'widgetIncidentTTP',
-    perspective: 'entities',
-    dataSelection: [
-      {
-        perspective: 'entities',
-        filters: {
-          mode: 'and',
-          filters: [
-            { key: 'entity_type', values: ['Indicator'] },
-            { key: 'objects', values: ['CONTAINER_ID'] },
-          ],
-          filterGroups: [],
-        },
-      },
-    ],
-  },
-};
-
-export default widgetIncidentTTP;
