--- opencti-platform/opencti-front/src/utils/outcome_template/widgets/incidentType.widget.ts
@@ -1,13 +1,0 @@
-import type { TemplateWidget } from '../template';
-
-const widgetIncidentType: TemplateWidget = {
-  name: 'Incident Type',
-  widget: {
-    type: 'attribute',
-    id: 'widgetIncidentType',
-    perspective: null,
-    dataSelection: [],
-  },
-};
-
-export default widgetIncidentType;
