+++ opencti-platform/opencti-graphql/config/schema/opencti.graphql
@@ -866,8 +866,9 @@ type TemplateWidget {
   widget: String!
 }
 type ResolvedWidgetAttribute {
-  template_widget_name: String
-  data: String
+  template_widget_name: String!
+  data: [String!]!
+  displayStyle: String
 }
 type TemplateAndUtils {
   template: Template!,
@@ -3116,6 +3117,7 @@ interface Container {
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
@@ -3301,6 +3303,7 @@ type Note implements BasicObject & StixObject & StixCoreObject & StixDomainObjec
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
@@ -3532,6 +3535,7 @@ type ObservedData implements BasicObject & StixObject & StixCoreObject & StixDom
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
@@ -3739,6 +3743,7 @@ type Opinion implements BasicObject & StixObject & StixCoreObject & StixDomainOb
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
@@ -3970,6 +3975,7 @@ type Report implements BasicObject & StixObject & StixCoreObject & StixDomainObj
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
