+++ opencti-platform/opencti-graphql/src/modules/case/case-incident/case-incident.graphql
@@ -146,6 +146,7 @@ type CaseIncident implements BasicObject & StixObject & StixCoreObject & StixDom
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
