+++ opencti-platform/opencti-graphql/src/modules/case/case-rfi/case-rfi.graphql
@@ -146,6 +146,7 @@ type CaseRfi implements BasicObject & StixObject & StixCoreObject & StixDomainOb
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
