+++ opencti-platform/opencti-graphql/src/modules/case/case.graphql
@@ -147,6 +147,7 @@ interface Case implements BasicObject & StixObject & StixCoreObject & StixDomain
     search: String
     filters: FilterGroup
   ): FileConnection
+  templates: [Template!]!
   templateAndUtils(
     templateId: String!
   ): TemplateAndUtils
