+++ opencti-platform/opencti-graphql/src/modules/grouping/grouping.graphql
@@ -154,6 +154,7 @@ type Grouping implements BasicObject & StixObject & StixCoreObject & StixDomainO
         search: String
         filters: FilterGroup
     ): FileConnection
+    templates: [Template!]!
     templateAndUtils(
         templateId: String!
     ): TemplateAndUtils
