+++ opencti-platform/opencti-graphql/src/resolvers/container.js
@@ -9,7 +9,9 @@ import {
   containersNumberByAuthor,
   containersNumberByEntity,
   containerEditAuthorizedMembers,
-  getContentsFromTemplate, getTemplateAndUtils
+  getContentsFromTemplate,
+  getTemplateAndUtils,
+  getTemplates
 } from '../domain/container';
 import {
   stixDomainObjectAddRelation,
@@ -50,6 +52,7 @@ const containerResolvers = {
     objects: (container, args, context) => objects(context, context.user, container.id, args),
     relatedContainers: (container, args, context) => relatedContainers(context, context.user, container.id, args),
     contentsFromTemplate: (container, { first, prefixMimeType }, context) => getContentsFromTemplate(context, context.user, container, { first, prefixMimeType }),
+    templates: (container, _, context) => getTemplates(context, context.user, container.id),
     templateAndUtils: (container, args, context) => getTemplateAndUtils(context, context.user, container.id, args.templateId),
   },
   // TODO Reactivate after official release of graphQL 17
